Implemeting comment replies. Closes #41.

+++ forum-api/src/main/java/com/github/mpacala00/forum/service/data/CommentServiceImpl.java
@@ -2,19 +2,21 @@ package com.github.mpacala00.forum.service.data;
 
 import com.github.mpacala00.forum.model.Comment;
 import com.github.mpacala00.forum.model.Post;
+import com.github.mpacala00.forum.model.constant.CommentConstants;
+import com.github.mpacala00.forum.model.dto.comment.CommentDTO;
 import com.github.mpacala00.forum.model.dto.comment.CommentUpdateDTO;
 import com.github.mpacala00.forum.repository.CommentRepository;
 import com.github.mpacala00.forum.repository.PostRepository;
 import com.github.mpacala00.forum.repository.UserRepository;
+import com.github.mpacala00.forum.service.dto.CommentDTOMappingService;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.util.CollectionUtils;
 
 import javax.transaction.Transactional;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.Optional;
-import java.util.Set;
+import java.util.*;
+import java.util.stream.Collectors;
 
 @Service
 @AllArgsConstructor
@@ -22,11 +24,21 @@ public class CommentServiceImpl implements CommentService {
 
     private final CommentRepository commentRepository;
     private final PostRepository postRepository;
+    private final CommentDTOMappingService commentDTOMappingService;
 
     @Override
-    public Set<Comment> getAllCommentsFromPost(Long postId) {
-        Optional<Post> post = postRepository.findById(postId);
-        return post.map(value -> new HashSet<>(value.getComments())).orElse(null);
+    public List<CommentDTO> getAllCommentsFromPost(Long postId) {
+        List<Comment> postComments = commentRepository.findByPostId(postId);
+        if (postComments.size() == 0) {
+            return Collections.emptyList();
+        }
+
+        List<CommentDTO> postCommentDTOList = postComments.stream()
+                .filter(comment -> comment.getParentComment() == null)
+                .map(commentDTOMappingService::convertToDTO)
+                .collect(Collectors.toList());
+
+        return postCommentDTOList;
     }
 
     @Override
@@ -59,10 +71,20 @@ public class CommentServiceImpl implements CommentService {
     @Transactional
     public void deleteById(Long id) {
         Comment commentToDelete = findById(id);
+        if (commentToDelete.getDeleted()) {
+            return;
+        }
 
-        commentToDelete.setPost(null);
-        commentToDelete.setCreator(null);
+        if (CollectionUtils.isEmpty(commentToDelete.getChildComments())) {
+            commentToDelete.setPost(null);
+            commentToDelete.setCreator(null);
 
-        commentRepository.deleteById(id);
+            commentRepository.deleteById(id);
+        } else {
+            commentToDelete.setBody(CommentConstants.DELETED_BODY.getValue());
+            commentToDelete.setDeleted(true);
+
+            commentRepository.save(commentToDelete);
+        }
     }
 }
